prepare v0.2.0
