+++ lib/embulk/output/hdfs.rb
@@ -1,3 +1,3 @@
 Embulk::JavaPlugin.register_output(
-  "hdfs", "org.embulk.output.HdfsOutputPlugin",
+  "hdfs", "org.embulk.output.hdfs.HdfsFileOutputPlugin",
   File.expand_path('../../../../classpath', __FILE__))
