--- src/test/java/org/embulk/output/TestHdfsOutputPlugin.java
@@ -1,5 +1,0 @@
-package org.embulk.output;
-
-public class TestHdfsOutputPlugin
-{
-}
