+++ src/test/java/org/embulk/output/hdfs/TestHdfsFileOutputPlugin.java
@@ -1,0 +1,5 @@
+package org.embulk.output.hdfs;
+
+public class TestHdfsFileOutputPlugin
+{
+}
